+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Version changes are pinned to SDK releases.
 - error: Move error related functionality to error.ts. ([#6](https://github.com/zetamarkets/sdk/pull/6))
 - insurance-client: Insurance fund functionality, whitelist checks, deposit & withdrawal. ([#9](https://github.com/zetamarkets/sdk/pull/9))
 - exchange: Insurance functionality, whitelist a user, rebalance vaults. ([#9](https://github.com/zetamarkets/sdk/pull/9))
+- general: Add functionality to SDK to support on chain options pricing. ([#11](https://github.com/zetamarkets/sdk/pull/11))
+- anchor: Bump to Anchor v0.18.2 typescript. ([#11](https://github.com/zetamarkets/sdk/pull/11))
 
 ## [0.7.3] 2021-11-29